+++ tests/panels/test_history.py
@@ -1,9 +1,9 @@
-from unittest.mock import patch
+import html
 
 from django.test import RequestFactory, override_settings
 from django.urls import resolve, reverse
 
-from debug_toolbar.panels.history.forms import HistoryStoreForm
+from debug_toolbar.forms import SignedDataForm
 from debug_toolbar.toolbar import DebugToolbar
 
 from ..base import BaseTestCase, IntegrationTestCase
@@ -66,6 +66,21 @@ class HistoryPanelTestCase(BaseTestCase):
 
 @override_settings(DEBUG=True)
 class HistoryViewsTestCase(IntegrationTestCase):
+    PANEL_KEYS = {
+        "VersionsPanel",
+        "TimerPanel",
+        "SettingsPanel",
+        "HeadersPanel",
+        "RequestPanel",
+        "SQLPanel",
+        "StaticFilesPanel",
+        "TemplatesPanel",
+        "CachePanel",
+        "SignalsPanel",
+        "LoggingPanel",
+        "ProfilingPanel",
+    }
+
     def test_history_panel_integration_content(self):
         """Verify the history panel's content renders properly.."""
         self.assertEqual(len(DebugToolbar._store), 0)
@@ -83,75 +98,76 @@ class HistoryViewsTestCase(IntegrationTestCase):
         response = self.client.get(reverse("djdt:history_sidebar"))
         self.assertEqual(response.status_code, 400)
 
-        data = {
-            "store_id": "foo",
-            "hash": "invalid",
-        }
+        data = {"signed": SignedDataForm.sign({"store_id": "foo"}) + "invalid"}
         response = self.client.get(reverse("djdt:history_sidebar"), data=data)
         self.assertEqual(response.status_code, 400)
-
-    @patch("debug_toolbar.panels.history.views.DebugToolbar.fetch")
-    def test_history_sidebar_hash(self, fetch):
-        """Validate the hashing mechanism."""
-        fetch.return_value.panels = []
-        data = {
-            "store_id": "foo",
-            "hash": "3280d66a3cca10098a44907c5a1fd255265eed31",
-        }
-        response = self.client.get(reverse("djdt:history_sidebar"), data=data)
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.json(), {})
 
     def test_history_sidebar(self):
         """Validate the history sidebar view."""
         self.client.get("/json_view/")
-        store_id = list(DebugToolbar._store.keys())[0]
-        data = {
-            "store_id": store_id,
-            "hash": HistoryStoreForm.make_hash({"store_id": store_id}),
-        }
+        store_id = list(DebugToolbar._store)[0]
+        data = {"signed": SignedDataForm.sign({"store_id": store_id})}
         response = self.client.get(reverse("djdt:history_sidebar"), data=data)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(
-            set(response.json().keys()),
-            {
-                "VersionsPanel",
-                "TimerPanel",
-                "SettingsPanel",
-                "HeadersPanel",
-                "RequestPanel",
-                "SQLPanel",
-                "StaticFilesPanel",
-                "TemplatesPanel",
-                "CachePanel",
-                "SignalsPanel",
-                "LoggingPanel",
-                "ProfilingPanel",
-            },
+            set(response.json()),
+            self.PANEL_KEYS,
         )
 
-    def test_history_refresh_invalid(self):
+    @override_settings(
+        DEBUG_TOOLBAR_CONFIG={"RESULTS_CACHE_SIZE": 1, "RENDER_PANELS": False}
+    )
+    def test_history_sidebar_expired_store_id(self):
+        """Validate the history sidebar view."""
+        self.client.get("/json_view/")
+        store_id = list(DebugToolbar._store)[0]
+        data = {"signed": SignedDataForm.sign({"store_id": store_id})}
+        response = self.client.get(reverse("djdt:history_sidebar"), data=data)
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(
+            set(response.json()),
+            self.PANEL_KEYS,
+        )
+        self.client.get("/json_view/")
+
+        # Querying old store_id should return in empty response
+        data = {"signed": SignedDataForm.sign({"store_id": store_id})}
+        response = self.client.get(reverse("djdt:history_sidebar"), data=data)
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.json(), {})
+
+        # Querying with latest store_id
+        latest_store_id = list(DebugToolbar._store)[0]
+        data = {"signed": SignedDataForm.sign({"store_id": latest_store_id})}
+        response = self.client.get(reverse("djdt:history_sidebar"), data=data)
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(
+            set(response.json()),
+            self.PANEL_KEYS,
+        )
+
+    def test_history_refresh_invalid_signature(self):
         response = self.client.get(reverse("djdt:history_refresh"))
         self.assertEqual(response.status_code, 400)
 
-        data = {
-            "store_id": "foo",
-            "hash": "invalid",
-        }
+        data = {"signed": "eyJzdG9yZV9pZCI6ImZvbyIsImhhc2giOiI4YWFiMzIzZGZhODIyMW"}
         response = self.client.get(reverse("djdt:history_refresh"), data=data)
         self.assertEqual(response.status_code, 400)
+        self.assertEqual(b"Invalid signature", response.content)
 
     def test_history_refresh(self):
         """Verify refresh history response has request variables."""
         data = {"foo": "bar"}
         self.client.get("/json_view/", data, content_type="application/json")
-        data = {
-            "store_id": "foo",
-            "hash": "3280d66a3cca10098a44907c5a1fd255265eed31",
-        }
+        data = {"signed": SignedDataForm.sign({"store_id": "foo"})}
         response = self.client.get(reverse("djdt:history_refresh"), data=data)
         self.assertEqual(response.status_code, 200)
         data = response.json()
         self.assertEqual(len(data["requests"]), 1)
+
+        store_id = list(DebugToolbar._store)[0]
+        signature = SignedDataForm.sign({"store_id": store_id})
+        self.assertIn(html.escape(signature), data["requests"][0]["content"])
+
         for val in ["foo", "bar"]:
             self.assertIn(val, data["requests"][0]["content"])