+++ tests/panels/test_sql.py
@@ -1,5 +1,7 @@
 import datetime
+import os
 import unittest
+from unittest.mock import patch
 
 import django
 from django.contrib.auth.models import User
@@ -9,6 +11,7 @@ from django.db.utils import DatabaseError
 from django.shortcuts import render
 from django.test.utils import override_settings
 
+import debug_toolbar.panels.sql.tracking as sql_tracking
 from debug_toolbar import settings as dt_settings
 
 from ..base import BaseTestCase
@@ -59,6 +62,20 @@ class SQLPanelTestCase(BaseTestCase):
 
         # ensure query was logged
         self.assertEqual(len(self.panel._queries), 1)
+
+    @patch("debug_toolbar.panels.sql.tracking.state", wraps=sql_tracking.state)
+    def test_cursor_wrapper_singleton(self, mock_state):
+        list(User.objects.all())
+
+        # ensure that cursor wrapping is applied only once
+        self.assertEqual(mock_state.Wrapper.call_count, 1)
+
+    @patch("debug_toolbar.panels.sql.tracking.state", wraps=sql_tracking.state)
+    def test_chunked_cursor_wrapper_singleton(self, mock_state):
+        list(User.objects.all().iterator())
+
+        # ensure that cursor wrapping is applied only once
+        self.assertEqual(mock_state.Wrapper.call_count, 1)
 
     def test_generate_server_timing(self):
         self.assertEqual(len(self.panel._queries), 0)
@@ -396,3 +413,47 @@ class SQLPanelTestCase(BaseTestCase):
         self.panel.generate_stats(self.request, response)
         self.assertEqual(len(self.panel._queries), 1)
         self.assertEqual(pretty_sql, self.panel._queries[-1][1]["sql"])
+
+    @override_settings(
+        DEBUG=True,
+    )
+    def test_flat_template_information(self):
+        """
+        Test case for when the query is used in a flat template hierarchy
+        (without included templates).
+        """
+        self.assertEqual(len(self.panel._queries), 0)
+
+        users = User.objects.all()
+        render(self.request, "sql/flat.html", {"users": users})
+
+        self.assertEqual(len(self.panel._queries), 1)
+
+        query = self.panel._queries[0]
+        template_info = query[1]["template_info"]
+        template_name = os.path.basename(template_info["name"])
+        self.assertEqual(template_name, "flat.html")
+        self.assertEqual(template_info["context"][2]["content"].strip(), "{{ users }}")
+        self.assertEqual(template_info["context"][2]["highlight"], True)
+
+    @override_settings(
+        DEBUG=True,
+    )
+    def test_nested_template_information(self):
+        """
+        Test case for when the query is used in a nested template
+        hierarchy (with included templates).
+        """
+        self.assertEqual(len(self.panel._queries), 0)
+
+        users = User.objects.all()
+        render(self.request, "sql/nested.html", {"users": users})
+
+        self.assertEqual(len(self.panel._queries), 1)
+
+        query = self.panel._queries[0]
+        template_info = query[1]["template_info"]
+        template_name = os.path.basename(template_info["name"])
+        self.assertEqual(template_name, "included.html")
+        self.assertEqual(template_info["context"][0]["content"].strip(), "{{ users }}")
+        self.assertEqual(template_info["context"][0]["highlight"], True)