+++ tests/test_integration.py
@@ -12,6 +12,7 @@ from django.template.loader import get_template
 from django.test import RequestFactory
 from django.test.utils import override_settings
 
+from debug_toolbar.forms import SignedDataForm
 from debug_toolbar.middleware import DebugToolbarMiddleware, show_toolbar
 from debug_toolbar.panels import Panel
 from debug_toolbar.toolbar import DebugToolbar
@@ -100,6 +101,25 @@ class DebugToolbarTestCase(BaseTestCase):
         self.assertEqual(len(self.toolbar.get_panel_by_id("CachePanel").calls), 3)
         self.client.get("/cached_view/")
         self.assertEqual(len(self.toolbar.get_panel_by_id("CachePanel").calls), 5)
+
+    def test_is_toolbar_request(self):
+        self.request.path = "/__debug__/render_panel/"
+        self.assertTrue(self.toolbar.is_toolbar_request(self.request))
+
+        self.request.path = "/invalid/__debug__/render_panel/"
+        self.assertFalse(self.toolbar.is_toolbar_request(self.request))
+
+        self.request.path = "/render_panel/"
+        self.assertFalse(self.toolbar.is_toolbar_request(self.request))
+
+    @override_settings(ROOT_URLCONF="tests.urls_invalid")
+    def test_is_toolbar_request_without_djdt_urls(self):
+        """Test cases when the toolbar urls aren't configured."""
+        self.request.path = "/__debug__/render_panel/"
+        self.assertFalse(self.toolbar.is_toolbar_request(self.request))
+
+        self.request.path = "/render_panel/"
+        self.assertFalse(self.toolbar.is_toolbar_request(self.request))
 
 
 @override_settings(DEBUG=True)
@@ -193,12 +213,15 @@ class DebugToolbarIntegrationTestCase(IntegrationTestCase):
     def test_sql_select_checks_show_toolbar(self):
         url = "/__debug__/sql_select/"
         data = {
-            "sql": "SELECT * FROM auth_user",
-            "raw_sql": "SELECT * FROM auth_user",
-            "params": "{}",
-            "alias": "default",
-            "duration": "0",
-            "hash": "6e12daa636b8c9a8be993307135458f90a877606",
+            "signed": SignedDataForm.sign(
+                {
+                    "sql": "SELECT * FROM auth_user",
+                    "raw_sql": "SELECT * FROM auth_user",
+                    "params": "{}",
+                    "alias": "default",
+                    "duration": "0",
+                }
+            )
         }
 
         response = self.client.post(url, data)
@@ -216,12 +239,15 @@ class DebugToolbarIntegrationTestCase(IntegrationTestCase):
     def test_sql_explain_checks_show_toolbar(self):
         url = "/__debug__/sql_explain/"
         data = {
-            "sql": "SELECT * FROM auth_user",
-            "raw_sql": "SELECT * FROM auth_user",
-            "params": "{}",
-            "alias": "default",
-            "duration": "0",
-            "hash": "6e12daa636b8c9a8be993307135458f90a877606",
+            "signed": SignedDataForm.sign(
+                {
+                    "sql": "SELECT * FROM auth_user",
+                    "raw_sql": "SELECT * FROM auth_user",
+                    "params": "{}",
+                    "alias": "default",
+                    "duration": "0",
+                }
+            )
         }
 
         response = self.client.post(url, data)
@@ -246,12 +272,15 @@ class DebugToolbarIntegrationTestCase(IntegrationTestCase):
         )
         query = base_query + """ '{"foo": "bar"}'"""
         data = {
-            "sql": query,
-            "raw_sql": base_query + " %s",
-            "params": '["{\\"foo\\": \\"bar\\"}"]',
-            "alias": "default",
-            "duration": "0",
-            "hash": "2b7172eb2ac8e2a8d6f742f8a28342046e0d00ba",
+            "signed": SignedDataForm.sign(
+                {
+                    "sql": query,
+                    "raw_sql": base_query + " %s",
+                    "params": '["{\\"foo\\": \\"bar\\"}"]',
+                    "alias": "default",
+                    "duration": "0",
+                }
+            )
         }
         response = self.client.post(url, data)
         self.assertEqual(response.status_code, 200)
@@ -268,12 +297,15 @@ class DebugToolbarIntegrationTestCase(IntegrationTestCase):
     def test_sql_profile_checks_show_toolbar(self):
         url = "/__debug__/sql_profile/"
         data = {
-            "sql": "SELECT * FROM auth_user",
-            "raw_sql": "SELECT * FROM auth_user",
-            "params": "{}",
-            "alias": "default",
-            "duration": "0",
-            "hash": "6e12daa636b8c9a8be993307135458f90a877606",
+            "signed": SignedDataForm.sign(
+                {
+                    "sql": "SELECT * FROM auth_user",
+                    "raw_sql": "SELECT * FROM auth_user",
+                    "params": "{}",
+                    "alias": "default",
+                    "duration": "0",
+                }
+            )
         }
 
         response = self.client.post(url, data)