+++ docs/changes.rst
@@ -4,17 +4,32 @@ Change log
 Next version
 ------------
 
-* Changed ``djdt.cookie.set()`` to set ``sameSite=Lax`` by default if
-  callers do not provide a value.
-* Added ``PRETTIFY_SQL`` configuration option to support controlling
-  SQL token grouping. By default it's set to True. When set to False,
-  a performance improvement can be seen by the SQL panel.
+* Ensured that the handle stays within bounds when resizing the window.
 * Disabled ``HistoryPanel`` when ``RENDER_PANELS`` is ``True``
   or if ``RENDER_PANELS`` is ``None`` and the WSGI container is
   running with multiple processes.
 * Fixed ``RENDER_PANELS`` functionality so that when ``True`` panels are
   rendered during the request and not loaded asynchronously.
 
+
+3.2.1 (2021-04-14)
+------------------
+
+* Fixed SQL Injection vulnerability, CVE-2021-30459. The toolbar now
+  calculates a signature on all fields for the SQL select, explain,
+  and analyze forms.
+* Changed ``djdt.cookie.set()`` to set ``sameSite=Lax`` by default if
+  callers do not provide a value.
+* Added ``PRETTIFY_SQL`` configuration option to support controlling
+  SQL token grouping. By default it's set to True. When set to False,
+  a performance improvement can be seen by the SQL panel.
+* Added a JavaScript event when a panel loads of the format
+  ``djdt.panel.[PanelId]`` where PanelId is the ``panel_id`` property
+  of the panel's Python class. Listening for this event corrects the bug
+  in the Timer Panel in which it didn't insert the browser timings
+  after switching requests in the History Panel.
+* Fixed issue with the toolbar expecting URL paths to start with
+  ``/__debug__/`` while the documentation indicates it's not required.
 
 3.2 (2020-12-03)
 ----------------